Skip size-limit on PRs from forks for now
The size-limit action requires an auth token to run, but this is not sent to runs triggered by PRs from forks, causing the CI to error. It’s not a very useful or friendly behaviour, so this commit skips size-limit entirely on PRs from forks.

TODO: Replace the size-limit action with something that uses GitHub’s annotations API instead so it can safely report without needing the token