Move sorting to sidebar logic to handle number-like routes
During auto-generation of sidebar groups, routes are turned into an object tree, which means we can lose ordering for some keys in particular a numbered key like `123`. This refactor moves that logic to the sidebar logic where it belongs and also paves the way to more easily using other entry data in sidebar generation.