Tabs: Remove margin-bottom being used to hide border.
We were setting margin-bottom as -2px when at the same time we had a
bottom border of 2px. Using this indirect margin was causing the tab
container to be greater in height than it's wrapper when we zoomed out
from the default zoom level. This introduced an unnecessary scroll bar.

We use a border-bottom of 2px on `.tab` mainly to mark the active tab
with `--sl-color-text-accent`, the additional margin-bottom: -2px hides
the grey border from the tablist so it doesn't feel like we have a blue
bottom border for the active tab sitting on top of a grey border.

Using box shadow instead to paint over the bottom border of the tablist
is a cleaner idea than doing these margin-bottom shenanigans.

Co-authored-by: Chris Swithinbank <swithinbank@gmail.com>